Re-design of signing mechanism
To make the verification process simpler and safer the signature and
signature meta data is now included directly in the header. The
signature is based only on the header and an additional hash
'payload_hash' is included in the header.

This way very few opertions must be performed to validate the integrity
of the header and at a later stage the computed payload hash can be
safely compared to the header payload hash.