Support for locating the header at the end of a package
Having the header at the end of a file or block device is necessary for most
 filsystems since linux can't deal with the filesystem starting at an offset.

This commit adds support for the alternative location of the header and includes
relevant updates for all of the currently supported algorithms.

It changes the behaviour of the transport sub-commands encode/decode by making
 '--output' a required parameter. Before the input file was updated during
encoding/decoding and this works poorly when dealing with block devices and
alternative header location.

New test case: test_transport2
Duplication of 'test_transport' with the differance that 'origin' data has
it's header stored at the end.

New test case: test_transport3:
Duplication of 'test_transport' with the differance that 'origin' data
and output has it's header stored at the end.

The transport --decode has an optional extra parameter '--output-header-last'
 that will write the bpak header at the last 4k in the output stream.