Add part hashes to header even with pre-computed signature file
One-step signing:
1. bpak sign archive.bpak --key privkey.pem

Part hashes are calculated and added to the header, which is then
hashed. The modified header is saved to the bpak file.

Two-step signing:
1. bpak show archive.bpak --hash > hash
2. <sign hash>
3. bpak sign archive.bpak --signature hash.signed

Part hashes are calculated in step 1 and added to the header in RAM
before hashing. This commit makes bpak calculate the part hashes in
step 3 as well and write them to the header in the file.