lib/transport_decode: Check BPAK_FLAG_TRANSPORT
Do not assume that the data is coded for transport just because there is
transport meta data present in the archive.

Add test_transport4.sh to check this behaviour.

This is needed so that the transport_decode API can be used generally
for archives that are encoded and 'raw', to similify streaming use
cases.