lib/transport_encode: bufix: transport_copy
transport_copy used the input 'patch header' to calculate both offsets
in input and output stream, which is incorrect.

The output stream will contain coded parts where the size is different
compared to the input part.

Also pass the output header as an argument so the output offset can be
calculated correctly.