python: Use setuptools for python installation
CMakes python/find_package does not work well in all environments. Using
python's setuptools seems to be the only way forward.

This commit removes the 'BPAK_BUILD_PYTHON_WRAPPER' option but keeps the
CMakefile under python/ for testing. This way the extension is built by
cmake when running tests and we can link against the locally built
bpak.so.